relocation of setup_train_loop
