+++ MaxText/elastic_train.py
@@ -71,7 +71,7 @@ from MaxText import pyconfig
 from MaxText.data_loader import DataLoader
 from MaxText.metric_logger import MetricLogger
 from MaxText.train import get_first_step
-from MaxText.train import setup_train_loop
+from MaxText.train_utils import setup_train_loop
 from MaxText.train import train_step
 from MaxText.train import validate_train_config
 from MaxText.utils.goodput_utils import (
