+++ MaxText/standalone_dataloader.py
@@ -30,7 +30,8 @@ import jax
 from MaxText import max_logging
 from MaxText import pyconfig
 from MaxText.data_loader import DataLoader
-from MaxText.train import validate_train_config, get_first_step, setup_train_loop
+from MaxText.train import validate_train_config, get_first_step
+from MaxText.train_utils import setup_train_loop
 
 
 def data_load_loop(config, state=None):
