+++ MaxText/train.py
@@ -131,10 +131,6 @@ def _split_dpo_state(state):
   return new_state, reference_params
 
 
-def _merge_dpo_state(state, reference_params):
-  return state.replace(params=dict(state.params, reference_params=reference_params))
-
-
 def dpo_loss_fn(model, config, data, dropout_rng, params, reference_params, is_train=True):
   """loss_fn for both train and eval.
 
@@ -305,19 +301,7 @@ def loss_fn(model, config, data, dropout_rng, params, is_train=True):
   xent = xent * (data["targets_segmentation"] != 0)
   total_loss = jnp.sum(xent)
   total_weights = jnp.sum(data["targets_segmentation"] != 0)
-
-  # If gradient accumulation is enabled, we don't need to divide total_loss
-  # by total_weights and then multiply the computed gradient by total_weights,
-  # since it's equivalent to computing the gradient from total_loss.
-  # This simplification reduces the number of operations and makes it easier
-  # for XLA to move all-reduce out of the gradient accumulation loop when use
-  # Zero1+GA to reduce communication overhead.
-  # EPS was used to avoid division by zero, but it's not needed when gradient
-  # accumulation is enabled since there's no division.
-  if config.gradient_accumulation_steps > 1:
-    loss = total_loss
-  else:
-    loss = total_loss / (total_weights + EPS)
+  loss = total_loss / (total_weights + EPS)
 
   # Calculate and Add MTP Loss
   mtp_loss = 0.0
@@ -380,7 +364,7 @@ def train_step(model, config, state_mesh_shardings, state, data, dropout_rng):
       acc_grad_and_loss["moe_lb_loss"] += aux["moe_lb_loss"]
       acc_grad_and_loss["mtp_loss"] += aux["mtp_loss"]
       acc_grad_and_loss["grad"] = jax.tree_util.tree_map(
-          lambda x, y: x + y, cur_batch_gradient, acc_grad_and_loss["grad"]
+          lambda x, y: x * aux["total_weights"] + y, cur_batch_gradient, acc_grad_and_loss["grad"]
       )
       acc_grad_and_loss["total_weights"] += aux["total_weights"]
       return acc_grad_and_loss, aux
@@ -508,101 +492,6 @@ def eval_step(model, config, state, data, dropout_rng):
   return metrics
 
 
-def setup_train_loop(config, recorder, devices=None):
-  """Set up prerequisites for the training loop -
-      checkpoint_manager, PRNG keys, Mesh, Model and optimizer.
-      Set up data iterator and tokenizer, initialize the model.
-
-  Args:
-    config
-    recorder
-
-  Returns:
-    init_rng:
-    checkpoint_manager: Orbax checkpointer
-    state_mesh_annotations: the mesh annotations for the train state
-    model:
-    mesh:
-    learning_rate_schedule:
-    data_iterator:
-    state: the initialized train state
-  """
-
-  with maybe_record_goodput(recorder, GoodputEvent.TPU_INIT):
-    model = mt.from_pretrained(config, devices)
-    mesh = model.mesh
-    init_rng, checkpoint_manager, learning_rate_schedule, tx = train_utils.create_training_tools(config, model, mesh)
-
-  with maybe_record_goodput(recorder, GoodputEvent.TRAINING_PREPARATION):
-    data_iterator, eval_data_iterator = create_data_iterator(config, mesh)
-    context_parallel_size = config.context_parallel_size
-    # Check if context parallelism is being used with sequence packing
-    if context_parallel_size > 1 and config.packing and config.dataset_type != "synthetic":
-      raise ValueError(
-          "Context parallelism cannot be used with sequence packing except for synthetic data where packing is not applied. "
-          "Either disable sequence packing (set packing=False) or disable context parallelism. "
-          "Context parallelism with packing support will be added soon."
-      )
-
-    # Apply reordering wrapper to data iterators if context parallelism is enabled
-    with mesh:
-      if context_parallel_size > 1 and config.context_parallel_load_balance:
-        data_iterator = map(max_utils.get_reorder_callable(context_parallel_size), data_iterator)
-        if eval_data_iterator:
-          eval_data_iterator = map(max_utils.get_reorder_callable(context_parallel_size), eval_data_iterator)
-
-    state, _, state_mesh_shardings, data_iterator = maxtext_utils.setup_training_state(
-        model, data_iterator, tx, config, init_rng, mesh, checkpoint_manager
-    )
-
-    # TODO(aireenmei, hengtaoguo): support sharding in vit for multimodal
-    if not config.using_pipeline_parallelism and not config.use_multimodal:
-      # The vocab tensor(s) of shape [vocab, embed] (and transpose) are not sharded by stage
-      maxtext_utils.assert_params_sufficiently_sharded(state.params, mesh, config.sharding_tolerance)
-
-    if config.use_dpo:
-      abstract_state, _, _ = maxtext_utils.get_abstract_state(model, tx, config, init_rng, mesh, is_training=True)
-      max_logging.log(f"Restoring reference parameters for DPO from '{os.path.join(str(config.checkpoint_dir), str(0))}'")
-      try:
-        step0_restored, _ = checkpointing.load_state_if_possible(
-            checkpoint_manager,
-            data_iterator,
-            load_parameters_from_path="",
-            load_full_state_from_path="",
-            checkpoint_storage_concurrent_gb=config.checkpoint_storage_concurrent_gb,
-            abstract_unboxed_pre_state=abstract_state,
-            enable_single_replica_ckpt_restoring=False,
-            dataset_type=config.dataset_type,
-            step=0,
-            use_ocdbt=config.checkpoint_storage_use_ocdbt,
-            use_zarr3=config.checkpoint_storage_use_zarr3,
-            enable_orbax_v1=config.enable_orbax_v1,
-            checkpoint_conversion_fn=config.checkpoint_conversion_fn,
-            source_checkpoint_layout=config.source_checkpoint_layout,
-        )
-      except FileNotFoundError:
-        step0_restored = None
-      if step0_restored is not None:
-        reference_params = step0_restored["items"].params["params"]
-        state = _merge_dpo_state(state, reference_params)
-      else:
-        max_logging.log(
-            f"Could not restore reference parameters for DPO from '{os.path.join(str(config.checkpoint_dir), str(0))}'"
-        )
-
-  return (
-      init_rng,
-      checkpoint_manager,
-      state_mesh_shardings,
-      model,
-      mesh,
-      learning_rate_schedule,
-      data_iterator,
-      eval_data_iterator,
-      state,
-  )
-
-
 def train_loop(config, recorder, state=None):
   """Main Training loop."""
   (
@@ -615,7 +504,7 @@ def train_loop(config, recorder, state=None):
       data_iterator,
       eval_data_iterator,
       state,
-  ) = setup_train_loop(config, recorder)
+  ) = train_utils.setup_train_loop(config, recorder)
 
   if config.use_dpo:
     if "reference_params" not in state.params:
